3.1.0
This revision adds a major change to the way EMG syncs proxy, product
and public key data with Edge. The purpose of this change is to add
resilience to startup and enable some features of EMG to work offline
even when Edge might not be fully reachable.
EMG now has the capability to read the bootstrap config data from Redis
To do so, set edge_config.redisBasedConfigCache to 'true'.
If set to 'false', EMG will continue to read the bootstrap data from
edge

Also, Redis server configs have been consolidated under 'edgemicro' section.
This affects quota. If quota was configured to use redis, those configs
now need to be moved to the 'edgemicro' section.

Allow redisDb value in range or 0 - 15

Use env variable EDGEMICRO_REDIS_PASSWORD if set and Override
'edgemicro.redisPassword'.

Below is an example of yaml configuration:

edge_config:
  redisBasedConfigCache: true

edgemicro:
  redisHost: localhost
  redisPort: 6379
  redisDb: 0
  redisPassword: <password>